have finished with interfaces and them realizations

+++ src/main/com/learning/bank/ATM.java
@@ -2,11 +2,11 @@ package main.com.learning.bank;
 
 public abstract class ATM {
 
-    public String getCash(Account acc, long money) {
+    public String getCash(Account acc, double money) {
         String result = "";
         try {
             result = acc.getMoney(money);
-        }catch (IllegalArgumentException e) {
+        } catch (IllegalArgumentException e) {
             System.err.println(e.getMessage());
         }
         return result;
